Corrections for Sun Studio compiler warnings.
git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@1652 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ src/mpid/osu_ch3/src/ch3u_handle_connection.c
@@ -300,7 +300,7 @@ int MPIDI_CH3_PktHandler_Close( MPIDI_VC_t *vc, MPIDI_CH3_Pkt_t *pkt,
 	{
 	    /* FIXME: Debugging */
 	    if (vc->state != MPIDI_VC_STATE_ACTIVE) {
-		printf("Unexpected state %d in vc %x\n", vc->state, (int)vc);
+		printf("Unexpected state %d in vc %x\n", vc->state, (unsigned long)vc);
 		fflush(stdout);
 	    }
 	    MPIU_DBG_MSG_D(CH3_DISCONNECT,VERBOSE,
